==> notebooks/01_scrapping.ipynb <==
#! /usr/bin/env python3.6
import requests
import datetime
from bs4 import BeautifulSoup
import time
import pandas as pd
import re
import pickle

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.ui import WebDriverWait
import numpy as np
from IPython.display import clear_output
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC
import time

import sys
sys.path.append("../")
import src.soporteScrapping as scr
import warnings
warnings.filterwarnings('ignore')


new

country = 'Sicily'
rooms = 1
adults = 2
checkin = '2023-09-01'
checkout = '2023-09-08'

list_url = []
list_dates = []
for i in range(1,30):
    if (len(str(i)) == 1) & (len(str(i+1)) == 1):
        day_in = '0' + str(i)
        day_out = '0' + str(i+1)
    elif (len(str(i)) == 1) & (len(str(i+1)) == 2):
        day_in = '0' + str(i)
        day_out = str(i+1)
    else:
        day_in = str(i)
        day_out = str(i+1)
    url_original_test = scr.url(rooms, country, f'2023-09-{day_in}', f'2023-09-{day_out}', adults)
    list_dates.append(f'2023-09-{day_in}')
    list_url.append(url_original_test)
list_url[0]

almacen = {
    'url':[],
    'name':[],
    'pagina':[],
    'address':[],
    'currentPrice':[],
    'previousPrice':[],
    'rating':[],
    'reviewNumber':[],
    'checkingDate':[],
    'image':[]
}

hace solo una paorte del bucle

scr.opcionesSelenium()

scr.seleniumChromeBooking_1st(list_url, list_dates, 0, almacen)

hace el bucle completo

scr.opcionesSelenium()

faltan registros día 0

dia = 3

chrome driver

scr.seleniumChromeBooking(list_url, list_dates, 0, almacen)

firefox driver

item29 page 4 day 3

item: 35 page: 3 day 5

item: 37 page: 3 day 0

scr.seleniumFirefoxBooking(list_url, list_dates, 0, almacen)
