==> notebooks/02_cleaning.ipynb <==
import pickle
import pandas as pd
import folium
import geopy
import geopandas
from tqdm import tqdm
import os
import warnings
warnings.filterwarnings("ignore")
import re

from geopy import Nominatim
locator = Nominatim(user_agent="myGeocoder")
location = locator.geocode("Champ de Mars, Paris, France")

# import data
I need to drop the `name` column cause there are some names missing.

os.chdir('..')
os.chdir('data')
os.chdir('pickle')
os.getcwd()

path = os.getcwd()
files = os.listdir(path)
files = [f for f in files if f[-6:] == 'pickle']
files

list_dicts = []
for file in files:
    with open(f'{file}', 'rb') as dict_booking:
        dict_booking = pickle.load(dict_booking)
        if len(dict_booking['url']) != len(dict_booking['name']):
            del dict_booking['url'][-1]
        print(len(dict_booking['url']), len(dict_booking['name']), len(dict_booking['pagina']), len(dict_booking['currentPrice']), len(dict_booking['address']))
        list_dicts.append(pd.DataFrame(dict_booking))

df_booking = pd.concat(list_dicts).reset_index(drop=True)
df_booking.head(1)

df_booking = df_booking.explode('address')
df_booking.head(1)

cleaning duplicated prices 

df_booking['currentPrice'].value_counts()

def countEuro(column):
    if column.count('€') == 1:
        return 1
    elif column.count('€') == 2:
        return 2
    else:
        return 0

df_booking['cleanPrice'] = df_booking['currentPrice'].apply(countEuro)

df_booking['cleanPrice'].value_counts()

df_booking_limpio = df_booking[df_booking['cleanPrice'] == 1]
df_booking_sucio = df_booking[df_booking['cleanPrice'] == 2]

df_booking_sucio.head(1)

df_booking_sucio[[0,1,2]] = df_booking_sucio['currentPrice'].str.split('€', expand=True, n = 2)
df_booking_sucio.head(1)

df_booking_sucio['currentPrice'] = '€ ' + df_booking_sucio[2] 
df_booking_sucio['previousPrice'] = '€ ' + df_booking_sucio[1]

df_booking_sucio.drop(['cleanPrice', 0, 1, 2], axis = 1, inplace=True)
df_booking_limpio.drop(['cleanPrice'], axis = 1, inplace=True)
df_booking_sucio.head(1)

df_booking = pd.concat([df_booking_limpio, df_booking_sucio])
print(df_booking.shape)
df_booking.head(1)

cleaning info -- parkings and other amenities
- comida
- ensuite bathroom
- piscina
- air conditioning

#parking
df_booking['parking'] = df_booking['pagina'].str.lower().str.findall(r'(free parking)')
df_booking = df_booking.explode('parking').drop_duplicates()
df_booking['parking'].value_counts()

#air conditioning
df_booking['air_conditioning'] = df_booking['pagina'].str.lower().str.findall(r'(air conditioning)')
df_booking = df_booking.explode('air_conditioning').drop_duplicates()
df_booking['air_conditioning'].value_counts()

#private bathroom
df_booking['bathroom'] = df_booking['pagina'].str.lower().str.findall(r'(private bathroom)')
df_booking = df_booking.explode('bathroom').drop_duplicates()
df_booking['bathroom'].value_counts()

#free wifi
df_booking['wifi'] = df_booking['pagina'].str.lower().str.findall(r'(free wifi)')
df_booking = df_booking.explode('wifi').drop_duplicates()
df_booking['wifi'].value_counts()

#breakfast
df_booking['breakfast'] = df_booking['pagina'].str.lower().str.findall(r'(breakfast)')
df_booking = df_booking.explode('breakfast').drop_duplicates()
df_booking['breakfast'].value_counts()

#coffee machine
df_booking['coffee_tea_maker'] = df_booking['pagina'].str.lower().str.findall(r'(tea/coffee maker)')
df_booking = df_booking.explode('coffee_tea_maker').drop_duplicates()
df_booking['coffee_tea_maker'].value_counts()

#laundry
df_booking['laundry'] = df_booking['pagina'].str.lower().str.findall(r'(laundry)')
df_booking = df_booking.explode('laundry').drop_duplicates()
df_booking['laundry'].value_counts()

df_booking.head(1)

print(df_booking.shape)
df_booking.drop_duplicates(inplace=True)
print(df_booking.shape)

checking dtypes

df_booking.dtypes

type(df_booking['address'][0])

# Geocoding

from geopy.extra.rate_limiter import RateLimiter
tqdm.pandas()

# 1 - conveneint function to delay between geocoding calls
geocode = RateLimiter(locator.geocode, min_delay_seconds=1)
# 2- - create location column
df_booking['location'] = df_booking['address'].progress_apply(geocode)
# 3 - create longitude, laatitude and altitude from location column (returns tuple)
df_booking['point'] = df_booking['location'].progress_apply(lambda loc: tuple(loc.point) if loc else None)
# 4 - split point column into latitude, longitude and altitude columns
df_booking[['latitude', 'longitude', 'altitude']] = pd.DataFrame(df_booking['point'].tolist(), index=df_booking.index)

map1 = folium.Map(
    location=[37.75120776952038, 14.992361985738308],
    tiles='cartodbpositron',
    zoom_start=12,
)
df_booking[df_booking['latitude'].isnull() == False].apply(lambda row:folium.CircleMarker(location=[row["latitude"], row["longitude"]]).add_to(map1), axis=1)
map1

df_booking.head(1)

df_booking.iloc[0,0]

revisa el formato de los precios, no está cogiendo bien los valores

df_booking['url_clean'] = df_booking['url'].str.split('?', expand=True)[0]
df_booking.head(1)

df_booking.drop(['url', 'pagina'], axis = 1).to_csv('../data/booking.csv', sep='|', encoding='utf-8', index = False)
