==> notebooks/03_istat_monthly_data_cleaning.ipynb <==
import numpy as np
import pandas as pd
pd.options.display.max_columns = None
import os

import sys
sys.path.append("../")
import src.biblioteca as bb
sys.path.append("../")
import src.soporteCleaning as scl

# data import

os.chdir('..')
os.chdir('data')
os.chdir('tourism_monthly')
os.getcwd()

path = os.getcwd()
files = os.listdir(path)
files = [f for f in files if f[-3:] == 'csv']
files

df = pd.read_csv(files[0], on_bad_lines='skip', sep = '|')
print(df.shape)
df.head()

# cleaning and adding comtrol columns

df['TIME'] = df['TIME'].str.replace('M', '-')

df['nivel'] = df['ITTER107'].apply(scl.nivelGeo)
df['nivel'].value_counts()

df['year_month'] = df['TIME'].apply(scl.yearOrMonthlyData)
df['year_month'].value_counts()

# monthly data

df_month = df[df['year_month'] == 'monthly']
print(df_month.shape)
df_month.head()

df_month['Type of accommodation'].value_counts()

df_month['nivel'].value_counts()

df_month['Indicators'].value_counts()

df_month['Country of residence of guests'].value_counts()

## arrivals and nights spent

df_month_arrivals = df_month[((df_month['Indicators'] == 'arrivals') |  
    (df_month['Indicators'] == 'nights spent')) & (df_month['Type of locality'] == 'all items') &
    ((df_month['Country of residence of guests'] == 'Italy') | 
        (df_month['Country of residence of guests'] == 'Foreign countries'))
    ]
print(df_month_arrivals.shape)
df_month_arrivals.head(1)

I will keep only the data available for Italy, Sicily region and provinces

df_month_arrivals = df_month_arrivals[df_month_arrivals['Territory'].isin(bb.list_region)]
print(df_month_arrivals.shape)
df_month_arrivals.head(1)

df_month_arrivals['Territory'].unique()

df_month_arrivals['Type of locality'].unique()

df_month_arrivals['Country of residence of guests'].value_counts()

df_month_arrivals['Country of residence of guests'].value_counts()

reorder columns & export

df_month_arrivals.head(1)

df_month_arrivals = df_month_arrivals[[df_month_arrivals.columns.tolist()[-2]] + df_month_arrivals.columns.tolist()[0:2] + [df_month_arrivals.columns.tolist()[3]] + [df_month_arrivals.columns.tolist()[7]] + [df_month_arrivals.columns.tolist()[-7]] + [df_month_arrivals.columns.tolist()[-5]]]
print(df_month_arrivals.shape)
df_month_arrivals.head(1)

df_month_arrivals.to_csv('../arrivals_nights_monthly.csv', sep = ';', decimal = '.')

## beds, baths and so on

df_month.head(1)

df_month['Indicators'].value_counts()

df_month_hotels = df_month[((df_month['Indicators'] == 'bed-places') | (df_month['Indicators'] == 'number of establishments') | (df_month['Indicators'] == 'bedrooms') | (df_month['Indicators'] == 'bathrooms')) & 
    (df_month['Coastal and non-coastal area'] == 'all items') & (df_month['Type of locality'] == 'all items')
    ]
print(df_month_hotels.shape)
df_month_hotels.head(1)

df_month_hotels['Degree of urbanisation'].value_counts()

df_month_hotels = df_month_hotels[(df_month_hotels['Territory'].isin(bb.list_region)) | ((len(df_month_hotels['ITTER107']) == 5) & (df_month_hotels['ITTER107'].str.startswith('8') == True)) ]
print(df_month_hotels.shape)
df_month_hotels.head(1)

df_month_hotels.to_csv('../hotels_monthly.csv', encoding='utf-8', sep=';', decimal='.', index = False)

# yearly data

df_year = df[df['year_month'] == 'yearly']
print(df_year.shape)
df_year.head()

df[df['TIME'] == '2021M1']

df['Type of accommodation'].value_counts()

alldf = []
for file in files:
    df = pd.read_csv(file)
    alldf.append(df)

df = pd.concat([alldf[0]])
print(df.shape)
df.head()

# clean values

df['nivel'] = df['ITTER107'].apply(scl.nivelGeo)
df['nivel'].value_counts()

df['period'] = df['TIME'].apply(scl.yearOrMonthlyData)
df['period'].value_counts()

df['Variable'] = df['Indicatori'].replace(bb.dict_indicatori)
df['Typology'] = df['Tipologia di esercizio'].replace(bb.dict_tipologia_esercizio)

df.head(1)

# reorder columns

lista_columnas = [df.columns.tolist()[-4]] + df.columns.tolist()[0:2] + df.columns.tolist()[-2:] + [df.columns.tolist()[-3]] + [df.columns.tolist()[-9]] + [df.columns.tolist()[-7]]
lista_columnas

new_df = df[lista_columnas]
new_df.head()

new_df.to_csv('../arrivals_nights_monthly.csv', encoding='utf-8', sep=';', decimal='.')
