==> notebooks/04_osm_pois.ipynb <==
import geopandas as gpd
import requests
import zipfile
import io
import matplotlib.pyplot as plt
import os
import fiona
import pandas as pd
import sys
#%matplotlib inline  # jupyter "magic" to display plots in notebook
pd.set_option('mode.chained_assignment', None)
import numpy as np
from bs4 import BeautifulSoup
import re
pd.options.display.max_columns = None
pd.options.mode.chained_assignment = None  # default='warn'
from geojson import Point, Feature, FeatureCollection, dump

import sys
sys.path.append("../")
import src.soporteScrapping as scr

parameters

url = 'https://download.geofabrik.de/europe/italy/'
country = 'IT'
country_clean = 'IT'

get links

links,flattened = scr.getLinks(url)
links

flattened

data conversion table input

ruta = os.path.dirname(os.getcwd()) + '\\data'

inPOIS = pd.ExcelFile(f'{ruta}\\POIS_Classification' + '.xlsx')
sheets_list = inPOIS.sheet_names[0]
POIS = pd.read_excel(inPOIS, sheet_name = sheets_list)
POIS.rename({'Tipo2':'Subtipo', 'Subtipo':'fclass'}, axis = 1, inplace=True)
POIS.head()

downloading shapefile

print('Downloading shapefile...')
patron = r'(.*).shp.zip'
for link, flat in zip(links, flattened):
    local_path = os.path.dirname(os.getcwd()) + f'\\data\\tmpOSM_{country}'
    r = requests.get(link)
    z = zipfile.ZipFile(io.BytesIO(r.content), mode='a')
    folder = re.findall(patron, flat)
    os.mkdir(os.path.join(local_path, folder[0]))
    z.extractall(path=(local_path + '//'+ folder[0])) # extract to folder
    print(f"Done with {link}")
filenames = [y for y in sorted(z.namelist()) for starting in ['gis_osm_landuse_a_free_1', 'gis_osm_buildings_a_free_1'] if y.startswith(starting)] 
print(filenames)

POIs

local_path = os.path.dirname(os.getcwd()) + f'\\data\\tmpOSM_{country}'

alldf_pois = []
for subdir, dirs, files in os.walk(local_path):
    #global alldf_landuse
    for file in files:
        #print(file)
        if file.startswith('gis_osm_pois_free_1.shp'):
            print(file, subdir)
            df_pois = gpd.read_file(subdir + '\\' + 'gis_osm_pois_free_1' + '.shp')
            alldf_pois.append(df_pois)

df_pois = pd.concat([(alldf_pois[n]) for n in list(range(len(alldf_pois)))])

print(df_pois.shape)
df_pois.head()

reproject

df_pois.crs ## tiene que estar en WGS 84

df_pois.to_crs("EPSG:4326", inplace=True)

df_pois.crs ## tiene que estar en WGS 84

spatial join geoLevels

GeoLevels_IT_03 = gpd.read_file(ruta + '\\' + 'GeoLevels_IT_03.geojson')
GeoLevels_IT_03.to_crs("EPSG:4326", inplace = True)

df_pois_merge = df_pois.sjoin(GeoLevels_IT_03, how="inner")
print(df_pois_merge.shape)
df_pois_merge.head(1)

merge with equivalence table

POIS.head()

df_pois_merge = df_pois_merge.merge(POIS, on ='fclass', how = 'left', indicator = True)
df_pois_merge['_merge'].value_counts()

df_pois_merge.info()

df_pois_merge.head()

selecting only the data that falls into sicily

df_pois_merge['nom_Region'].unique()

df_pois_merge = df_pois_merge[df_pois_merge['nom_Region'] == 'Sicilia']

df_pois_merge.drop(['_merge'], axis = 1, inplace = True)

add hotel listings to paint both of them together

local_path = os.path.dirname(os.getcwd()) + f'\\data'
local_path

df_booking = pd.read_csv(local_path + '//booking.csv', sep =';', decimal=',', thousands='.')
df_booking.head(1)

turn into geodataframe and remove unnecesary columns

df_booking = gpd.GeoDataFrame(df_booking, geometry=gpd.points_from_xy(df_booking.longitude, df_booking.latitude))
df_booking.set_crs("EPSG:4326", inplace = True)
df_booking.head(1)

df_pois_merge['currenPrice'] = 100
df_pois_merge.head(1)

df_booking['fclass'] = 'hotel'
df_booking['Tipo'] = 'Hotel'
df_booking['Subtipo'] = 'Hotel'
df_booking.drop(['address', 'previousPrice', 'rating', 'reviewNumber', 'image', 'parking', 'air_conditioning', 'bathroom', 'wifi', 'breakfast', 'coffee_tea_maker', 'laundry', 'location', 'point', 
                 'latitude', 'longitude', 'altitude', 'url_clean'], axis = 1, inplace = True)
df_booking.head(1)

df_total = pd.concat([df_pois_merge, df_booking])
print(df_total.shape)
df_total.head(1)

export

df_booking.dtypes

local_path = os.path.dirname(os.getcwd()) + f'\\data'
local_path

df_total['lon'] = df_total.geometry.apply(lambda p: p.x)
df_total['lat'] = df_total.geometry.apply(lambda p: p.y)
df_total.head(1)

df_total.to_file(f'{local_path}\\df_pois.geojson', driver='GeoJSON')  

df_total.head(1)

df_total['currentPrice'].value_counts()

df_total.to_csv(f'{local_path}//pois.csv', encoding='utf-8', sep='|', decimal='.', index=False)

print(df_total[df_pois_merge['nom_Region'] == 'Sicilia'].shape)
