==> notebooks/05_weatherAPI.ipynb <==
import pandas as pd
import requests

# api data

url = 'https://archive-api.open-meteo.com/v1/archive?latitude=37.62&longitude=13.93&start_date=2008-01-01&end_date=2023-02-04&hourly=temperature_2m,relativehumidity_2m&models=era5_land&timezone=Europe%2FBerlin'

results = requests.get(url)
results

# clean api data

df_tiempo = pd.DataFrame(results.json()['hourly'])
df_tiempo

df_tiempo.to_csv('../data/weather_sicily.csv', sep='|')
